% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-site.R
\name{build_ro_site}
\alias{build_ro_site}
\title{Build website using rotemplate}
\usage{
build_ro_site(pkg = ".", preview = TRUE, ...)
}
\arguments{
\item{pkg}{Path to package folder}

\item{preview}{Wether to preview the website (servr will be used)}

\item{...}{}
}
\value{
Nothing
}
\description{
Build website using rotemplate
}
