% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-site.R
\name{build_ro_site}
\alias{build_ro_site}
\title{Build website using rotemplate}
\usage{
build_ro_site(pkg = ".", preview = TRUE, override = NULL, ...)
}
\arguments{
\item{pkg}{Path to package folder}

\item{preview}{preview the website (servr will be used)}

\item{override}{passed to [pkgdown::build_site]}

\item{...}{}
}
\value{
Nothing
}
\description{
Build website using rotemplate
}
